feat(144151): Adiciona hooks para post de finalizar processo
